improving error message
